pyproject.toml
[project]
name = "a2a-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "sse-starlette>=3.1.2",
    "starlette>=0.50.0",
    "uvicorn>=0.40.0",
]
